requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.36
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6
asyncpg==0.29.0
Werkzeug==2.3.7
httpx==0.27.0
pydantic==2.11.7
anyio==3.7.1
click==8.1.7